Do not require bindings in IdP metadata parser
In some use cases, it is valuable to return the first binding
the IdP supports for SSO/SLO endpoints. Rather than calling
`OneLogin_Saml2_IdPMetadataParser.parse` with a binding,
testing to see if an endpoint was found, and repeating the process
until and endpoint is found, simply calling `.parse` and checking
what binding was returned may be more usable.

Only two bindings are tested for: REDIRECT and POST

This change set is backwards compatible.