Replace double-underscored names with single underscores
This will avoid Python's class-name-prefix mangling of attributes and methods
while still retaining the signal that they're meant for private use of the library.

This makes it easier to e.g. subclass classes should one need to, and to access
data from outside (while being mindful of the fact that you're accessing notionally
private data).

Only one change to a test was required (and that test was indeed accessing a private
field from outside...).